Integrate expo-router and extract overview screen
