+++ app/_layout.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Stack } from "expo-router";
+
+import { AuthGate } from "../src/components/AuthGate";
+
+export default function RootLayout() {
+  return (
+    <AuthGate>
+      <Stack screenOptions={{ headerShown: false }} />
+    </AuthGate>
+  );
+}
