+++ app/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import AuthenticatedApp from "../src/app/AuthenticatedApp";
+
+export default function Index() {
+  return <AuthenticatedApp />;
+}
