+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "aibarber",
   "version": "1.0.0",
-  "main": "index.ts",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "web-build": "expo export -p web",
@@ -14,6 +14,7 @@
     "@expo/metro-runtime": "~5.0.4",
     "@supabase/supabase-js": "^2.75.0",
     "expo": "~54.0.13",
+    "expo-router": "^4.0.7",
     "expo-status-bar": "~3.0.8",
     "expo-localization": "~17.0.7",
     "react": "19.0.0",
